% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT data set}
\format{
A data set with 144 rows and 5 variables:
\describe{
  \item{LENGTH}{length of fish}
  \item{WEIGHT}{weight of the fish}
  \item{SPECIES}{Species type of fish}
  ...
}
}
\source{
\url{Textbook: Statistics for Engineering and the Sciences}
}
\usage{
ddt
}
\description{
Data containing DDT in different fish species
}
\keyword{datasets}
