% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Sampling Function showing Binomial simulation}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{p}{}
}
\value{
Bar plot of successes
}
\description{
Sampling Function showing Binomial simulation
}
\examples{
mybin(iter=1000,n=18, p=0.3)
}
