% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{My clt Function}
\usage{
myclt(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}

\item{a}{a variable}

\item{b}{b variable}
}
\value{

}
\description{
My clt Function
}
\examples{
w=myclt(n=50,iter=10000,a=5,b=10)
}
