% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddt.R
\name{myddtfun2}
\alias{myddtfun2}
\title{myddt function}
\usage{
myddtfun2(df, x, y, cond, col)
}
\arguments{
\item{df}{data}

\item{x}{x variable (Length)}

\item{y}{y variable (Weight}

\item{cond}{length of species}

\item{col}{column variable}
}
\value{
Plots ddt data
}
\description{
myddt function
}
\examples{

}
