% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Curve Function}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean value}

\item{sigma}{standard deviation}

\item{a}{secondxcurvevalue}
}
\value{
Plots Curve
}
\description{
Curve Function
}
\examples{
myncurve(mu=10,sigma=5, a=6)
}
